requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=./requirements/requirements.txt ./requirements/requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via httpx
asgiref==3.7.2
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
deprecation==2.1.0
    # via postgrest
django==5.0.1
    # via
    #   -r ./requirements/requirements.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-filter==23.5
    # via -r ./requirements/requirements.in
djangorestframework==3.14.0
    # via
    #   -r ./requirements/requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.3.1
    # via -r ./requirements/requirements.in
drf-spectacular==0.27.1
    # via -r ./requirements/requirements.in
forex-python==1.8
    # via -r ./requirements/requirements.in
gotrue==2.1.0
    # via supabase
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.25.2
    # via
    #   gotrue
    #   postgrest
    #   storage3
    #   supabase
    #   supafunc
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown==3.5.2
    # via -r ./requirements/requirements.in
packaging==23.2
    # via deprecation
pillow==10.2.0
    # via -r ./requirements/requirements.in
postgrest==0.15.0
    # via supabase
pydantic==2.5.3
    # via
    #   gotrue
    #   postgrest
pydantic-core==2.14.6
    # via pydantic
pyjwt==2.8.0
    # via djangorestframework-simplejwt
python-dateutil==2.8.2
    # via
    #   realtime
    #   storage3
python-decouple==3.8
    # via -r ./requirements/requirements.in
pytz==2023.3.post1
    # via djangorestframework
pyyaml==6.0.1
    # via drf-spectacular
realtime==1.0.2
    # via supabase
referencing==0.32.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   forex-python
    #   stripe
rpds-py==0.17.1
    # via
    #   jsonschema
    #   referencing
segno==1.6.0
    # via -r ./requirements/requirements.in
simplejson==3.19.2
    # via forex-python
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
sqlparse==0.4.4
    # via django
storage3==0.7.0
    # via supabase
strenum==0.4.15
    # via postgrest
stripe==7.13.0
    # via -r ./requirements/requirements.in
supabase==2.3.4
    # via -r ./requirements/requirements.in
supafunc==0.3.3
    # via supabase
typing-extensions==4.9.0
    # via
    #   pydantic
    #   pydantic-core
    #   realtime
    #   storage3
    #   stripe
tzdata==2023.4
    # via django
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.1.0
    # via requests
websockets==11.0.3
    # via realtime
